% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_PSweight.R
\name{coef.PSweight}
\alias{coef.PSweight}
\title{Point estimates of PSweight}
\usage{
\method{coef}{PSweight}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model coefficients is meaningful.}

\item{...}{other arguments.}
}
\value{
The output from \code{\link{coef}}
}
\description{
The \code{\link{coef}} method for class "PSweight"
}
