% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PStrim.R
\name{print.PStrim}
\alias{print.PStrim}
\title{Print the results of PStrim}
\usage{
\method{print}{PStrim}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "PStrim"
}
