% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PSweight.R
\name{print.PSweight}
\alias{print.PSweight}
\title{Print the results of PSweight}
\usage{
\method{print}{PSweight}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "PSweight"
}
