% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PSweightsum.R
\name{print.PSweightsum}
\alias{print.PSweightsum}
\title{Print the results of Summary.PSweight}
\usage{
\method{print}{PSweightsum}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The output from \code{\link{print}}
}
\description{
The \code{\link{print}} method for class "PSweightsum"
}
