% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psdata.R
\docType{data}
\name{psdata}
\alias{psdata}
\title{Simulated dataset for PSweight}
\format{
A data frame with 1500 rows and 8 columns.
}
\usage{
data(psdata)
}
\description{
This is a simulated observational study with three treatment groups to illustated the ulity of PSweight.
}
\details{
The simulated dataset includes 1500 rows, with each row represents information recorded from each individual.
There are 8 variables (columns). The treatment is the variable trt, which has three treatment arms.
The outcome of interest is variable Y. cov1-cov6 are pre-treatment covariates among which cov1-cov5 are continous, and cov6 is binary.
}
\examples{
data("psdata")
}
\keyword{dataset}
